remote_sensing/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
aenum=3.1.15=pypi_0
affine=2.4.0=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
attrs=23.2.0=pypi_0
bitsandbytes=0.41.0=pypi_0
bzip2=1.0.8=hd590300_5
ca-certificates=2024.2.2=hbcca054_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.2=pypi_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
docstring-parser=0.16=pypi_0
efficientnet-pytorch=0.7.1=pypi_0
einops=0.7.0=pypi_0
featup=0.1.2=dev_0
filelock=3.13.1=pypi_0
fiona=1.9.6=pypi_0
fonttools=4.50.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
geopandas=0.14.3=pypi_0
huggingface-hub=0.21.4=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.6=pypi_0
importlib-resources=6.3.2=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jsonargparse=4.27.6=pypi_0
kiwisolver=1.4.5=pypi_0
kornia=0.7.2=pypi_0
kornia-rs=0.1.2=pypi_0
ld_impl_linux-64=2.40=h41732ed_0
libexpat=2.6.2=h59595ed_0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_5
libgomp=13.2.0=h807b86a_5
libnsl=2.0.1=hd590300_0
libsqlite=3.45.2=h2797004_0
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=hd590300_5
lightly=1.5.1=pypi_0
lightly-utils=0.0.2=pypi_0
lightning=2.2.1=pypi_0
lightning-utilities=0.11.0=pypi_0
llvmlite=0.42.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.3=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
munch=4.0.0=pypi_0
ncurses=6.4.20240210=h59595ed_0
networkx=3.2.1=pypi_0
numba=0.59.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.4.99=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
omegaconf=2.3.0=pypi_0
openssl=3.2.1=hd590300_1
packaging=24.0=pypi_0
pandas=2.2.1=pypi_0
pillow=10.2.0=pypi_0
pip=24.0=pyhd8ed1ab_0
pretrainedmodels=0.7.4=pypi_0
protobuf=5.26.0=pypi_0
pydantic=1.10.14=pypi_0
pygments=2.17.2=pypi_0
pynndescent=0.5.11=pypi_0
pyparsing=3.1.2=pypi_0
pyproj=3.6.1=pypi_0
python=3.12.2=hab00c5b_0_cpython
python-dateutil=2.9.0.post0=pypi_0
pytorch-lightning=2.2.1=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
rasterio=1.3.9=pypi_0
readline=8.2=h8228510_1
requests=2.31.0=pypi_0
rich=13.7.1=pypi_0
rtree=1.2.0=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.12.0=pypi_0
segmentation-models-pytorch=0.3.3=pypi_0
setuptools=69.2.0=pypi_0
shapely=2.0.3=pypi_0
six=1.16.0=pypi_0
snuggs=1.4.7=pypi_0
sympy=1.12=pypi_0
tensorboardx=2.6.2.2=pypi_0
threadpoolctl=3.4.0=pypi_0
timm=0.9.2=pypi_0
tk=8.6.13=noxft_h4845f30_101
torch=2.2.1=pypi_0
torchaudio=2.2.1=pypi_0
torchgeo=0.5.2=pypi_0
torchmetrics=1.3.2=pypi_0
torchvision=0.17.1=pypi_0
tqdm=4.66.2=pypi_0
typeshed-client=2.5.1=pypi_0
typing-extensions=4.10.0=pypi_0
tzdata=2024.1=pypi_0
umap-learn=0.5.5=pypi_0
urllib3=2.2.1=pypi_0
wheel=0.42.0=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
yarl=1.9.4=pypi_0
